# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DataTable MCP Server API",
    "description": "Model Context Protocol (MCP) server for in-memory tabular data manipulation. Provides 21 tools across 6 categories for comprehensive data operations including table lifecycle management, data manipulation, querying, export/import, advanced operations, and session management.",
    "version": "1.0.0",
    "contact": {
      "name": "DataTable MCP Server"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8321",
      "description": "HTTP transport mode for testing and web integration"
    },
    {
      "url": "stdio://mcp",
      "description": "Standard I/O transport (default MCP mode for Claude Desktop)"
    }
  ],
...
  "paths": {
    "/mcp/tools/create_table": {
      "post": {
        "tags": [
          "Table Lifecycle"
        ],
        "summary": "Create a new DataTable from data array with auto-detected headers.",
        "description": "Create a new DataTable from data array with auto-detected headers.\n\nArgs:\n    data: 2D array of table data (rows x columns)\n    headers: Optional column headers. If not provided, will auto-generate (Column_1, Column_2, etc.)\n    name: Optional table name for identification\n\nReturns:\n    Dict containing table_id and basic table information",
        "operationId": "create_table",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "data"
                ],
                "properties": {
                  "data": {
                    "description": "Parameter data",
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "headers": {
                    "description": "Parameter headers",
                    "type": "string"
                  },
                  "name": {
                    "description": "Parameter name",
                    "type": "string"
                  }
                },
                "example": {
                  "data": [
                    [
                      "Alice",
                      25,
                      "Engineer"
                    ],
                    [
                      "Bob",
                      30,
                      "Manager"
                    ]
                  ],
                  "headers": [
                    "Name",
                    "Age",
                    "Role"
                  ],
                  "name": "Employees"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful tool execution",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPToolResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/mcp/tools/load_table": {
      "post": {
        "tags": [
          "Table Lifecycle"
        ],
        "summary": "Load a table from external sources (spreadsheet/excel/database).",
        "description": "Load a table from external sources (spreadsheet/excel/database).\n\nArgs:\n    source_type: Type of source (\"google_sheets\", \"excel\", \"csv\", \"database\")\n    source_path: Path/URL to the source (file path, spreadsheet ID, database connection string)\n    name: Optional table name\n    sheet_name: Sheet name for Excel/Google Sheets (optional)\n    encoding: File encoding for CSV files (optional)\n    delimiter: Delimiter for CSV files (optional)\n\nReturns:\n    Dict containing table_id and loaded table information",
        "operationId": "load_table",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "source_type",
                  "source_path"
                ],
                "properties": {
                  "source_type": {
                    "description": "Parameter source_type",
                    "type": "string"
                  },
                  "source_path": {
                    "description": "Parameter source_path",
                    "type": "string"
                  },
                  "name": {
                    "description": "Parameter name",
                    "type": "string"
                  },
                  "sheet_name": {
                    "description": "Parameter sheet_name",
                    "type": "string"
                  },
                  "encoding": {
                    "description": "Parameter encoding",
                    "type": "string"
                  },
                  "delimiter": {
                    "description": "Parameter delimiter",
                    "type": "string"
                  }
                },
                "example": {
                  "table_id": "dt_abc12345"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful tool execution",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPToolResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/mcp/tools/clone_table": {
      "post": {
        "tags": [
          "Table Lifecycle"
        ],
        "summary": "Create a deep copy of an existing table.",
        "description": "Create a deep copy of an existing table.\n\nArgs:\n    source_table_id: ID of the table to clone\n    new_name: Optional name for the cloned table\n\nReturns:\n    Dict containing new table_id and cloned table information",
        "operationId": "clone_table",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "source_table_id"
                ],
                "properties": {
                  "source_table_id": {
                    "description": "Parameter source_table_id",
                    "type": "string"
                  },
                  "new_name": {
                    "description": "Parameter new_name",
                    "type": "string"
                  }
                },
                "example": {
                  "table_id": "dt_abc12345"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful tool execution",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPToolResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/mcp/tools/list_tables": {
      "post": {
        "tags": [
          "Table Lifecycle"
        ],
        "summary": "Get inventory of all tables in the current session.",
        "description": "Get inventory of all tables in the current session.\n\nReturns:\n    Dict containing list of all active tables with their basic information",
        "operationId": "list_tables",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [],
                "properties": {},
                "example": {
                  "table_id": "dt_abc12345"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful tool execution",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPToolResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/mcp/tools/append_row": {
      "post": {
        "tags": [
          "Data Manipulation"
        ],
        "summary": "Add a new row to the table with robust handling of dimension mismatches.",
        "description": "Add a new row to the table with robust handling of dimension mismatches.\n\nArgs:\n    table_id: ID of the target table\n    row_data: List of values for the new row\n    fill_strategy: How to handle dimension mismatches (\"none\", \"fill_na\", \"fill_empty\", \"fill_zero\")\n\nReturns:\n    Dict containing success status and updated table information",
        "operationId": "append_row",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "table_id",
                  "row_data"
                ],
                "properties": {
                  "table_id": {
                    "description": "Parameter table_id",
                    "type": "string"
                  },
                  "row_data": {
                    "description": "Parameter row_data",
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "fill_strategy": {
                    "description": "Parameter fill_strategy",
                    "type": "string"
                  }
                },
                "example": {
                  "table_id": "dt_abc12345"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful tool execution",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPToolResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/mcp/tools/add_column": {
      "post": {
        "tags": [
          "Data Manipulation"
        ],
        "summary": "Add a new column to the table with optional default values.",
        "description": "Add a new column to the table with optional default values.\n\nArgs:\n    table_id: ID of the target table\n    column_name: Name of the new column\n    default_value: Default value for all rows in the new column\n    position: Optional position to insert column (if not specified, adds at end)\n\nReturns:\n    Dict containing success status and updated table information",
        "operationId": "add_column",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "table_id",
                  "column_name"
                ],
                "properties": {
                  "table_id": {
                    "description": "Parameter table_id",
                    "type": "string"
                  },
                  "column_name": {
                    "description": "Parameter column_name",
                    "type": "string"
                  },
                  "default_value": {
                    "description": "Parameter default_value",
                    "type": "string"
                  },
                  "position": {
                    "description": "Parameter position",
                    "type": "string"
                  }
                },
                "example": {
                  "table_id": "dt_abc12345"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful tool execution",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPToolResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/mcp/tools/set_range_values": {
      "post": {
        "tags": [
          "Data Manipulation"
        ],
        "summary": "Update values in specified ranges using pandas .loc style updates with fill strategies.",
        "description": "Update values in specified ranges using pandas .loc style updates with fill strategies.\n\nArgs:\n    table_id: ID of the target table\n    row_indices: List of row indices to update\n    column_names: List of column names to update\n    values: 2D array of new values [rows x columns]\n    fill_strategy: How to handle dimension mismatches (\"none\", \"fill_na\", \"fill_empty\", \"fill_zero\")\n\nReturns:\n    Dict containing success status and update information",
        "operationId": "set_range_values",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "table_id",
                  "row_indices",
                  "column_names",
                  "values"
                ],
                "properties": {
                  "table_id": {
                    "description": "Parameter table_id",
                    "type": "string"
                  },
                  "row_indices": {
                    "description": "Parameter row_indices",
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "column_names": {
                    "description": "Parameter column_names",
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "values": {
                    "description": "Parameter values",
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "fill_strategy": {
                    "description": "Parameter fill_strategy",
                    "type": "string"
                  }
                },
                "example": {
                  "table_id": "dt_abc12345"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful tool execution",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPToolResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/mcp/tools/delete_from_table": {
      "post": {
        "tags": [
          "Data Manipulation"
        ],
        "summary": "Unified deletion of rows or columns from the table.",
        "description": "Unified deletion of rows or columns from the table.\n\nArgs:\n    table_id: ID of the target table\n    target_type: Type of deletion (\"rows\" or \"columns\")\n    target_names: List of row indices (for rows) or column names (for columns)\n\nReturns:\n    Dict containing success status and deletion information",
        "operationId": "delete_from_table",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "table_id",
                  "target_type",
                  "target_names"
                ],
                "properties": {
                  "table_id": {
                    "description": "Parameter table_id",
                    "type": "string"
                  },
                  "target_type": {
                    "description": "Parameter target_type",
                    "type": "string"
                  },
                  "target_names": {
                    "description": "Parameter target_names",
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "table_id": "dt_abc12345"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful tool execution",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPToolResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/mcp/tools/rename_columns": {
      "post": {
        "tags": [
          "Data Manipulation"
        ],
        "summary": "Rename multiple columns in bulk.",
        "description": "Rename multiple columns in bulk.\n\nArgs:\n    table_id: ID of the target table\n    column_mapping: Dictionary mapping old column names to new column names\n\nReturns:\n    Dict containing success status and rename information",
        "operationId": "rename_columns",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "table_id",
                  "column_mapping"
                ],
                "properties": {
                  "table_id": {
                    "description": "Parameter table_id",
                    "type": "string"
                  },
                  "column_mapping": {
                    "description": "Parameter column_mapping",
                    "type": "object"
                  }
                },
                "example": {
                  "table_id": "dt_abc12345"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful tool execution",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPToolResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/mcp/tools/clear_range": {
      "post": {
        "tags": [
          "Data Manipulation"
        ],
        "summary": "Clear values in specified range while preserving table structure.",
        "description": "Clear values in specified range while preserving table structure.\n\nArgs:\n    table_id: ID of the target table\n    row_indices: Optional list of row indices to clear (if None, clears all rows)\n    column_names: Optional list of column names to clear (if None, clears all columns)\n    clear_value: Value to set cleared cells to (default: None/NaN)\n\nReturns:\n    Dict containing success status and clear operation information",
        "operationId": "clear_range",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "table_id"
                ],
                "properties": {
                  "table_id": {
                    "description": "Parameter table_id",
                    "type": "string"
                  },
                  "row_indices": {
                    "description": "Parameter row_indices",
                    "type": "string"
                  },
                  "column_names": {
                    "description": "Parameter column_names",
                    "type": "string"
                  },
                  "clear_value": {
                    "description": "Parameter clear_value",
                    "type": "string"
                  }
                },
                "example": {
                  "table_id": "dt_abc12345"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful tool execution",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPToolResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/mcp/tools/get_table_data": {
      "post": {
        "tags": [
          "Data Query"
        ],
        "summary": "Get table data with flexible slicing and multiple output formats.",
        "description": "Get table data with flexible slicing and multiple output formats.\n\nArgs:\n    table_id: ID of the target table\n    output_format: Output format (\"dict\", \"records\", \"values\", \"json\")\n    start_row: Starting row index (inclusive)\n    end_row: Ending row index (exclusive)\n    columns: Optional list of specific columns to return\n    max_rows: Maximum number of rows to return (safety limit)\n\nReturns:\n    Dict containing table data in requested format",
        "operationId": "get_table_data",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "table_id"
                ],
                "properties": {
                  "table_id": {
                    "description": "Parameter table_id",
                    "type": "string"
                  },
                  "output_format": {
                    "description": "Parameter output_format",
                    "type": "string"
                  },
                  "start_row": {
                    "description": "Parameter start_row",
                    "type": "string"
                  },
                  "end_row": {
                    "description": "Parameter end_row",
                    "type": "string"
                  },
                  "columns": {
                    "description": "Parameter columns",
                    "type": "string"
                  },
                  "max_rows": {
                    "description": "Parameter max_rows",
                    "type": "string"
                  }
                },
                "example": {
                  "table_id": "dt_abc12345",
                  "format": "json"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful tool execution",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPToolResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/mcp/tools/sort_table": {
      "post": {
        "tags": [
          "Data Query"
        ],
        "summary": "Sort table by multiple columns with option to create new table or modify in-place.",
        "description": "Sort table by multiple columns with option to create new table or modify in-place.\n\nArgs:\n    table_id: ID of the target table\n    sort_columns: List of column names to sort by (in order of priority)\n    ascending: List of boolean values indicating sort direction for each column (True = ascending)\n    in_place: If True, modifies the original table; if False, creates a new table\n    new_table_name: Name for the new table (if in_place is False)\n\nReturns:\n    Dict containing sort operation information",
        "operationId": "sort_table",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "table_id",
                  "sort_columns"
                ],
                "properties": {
                  "table_id": {
                    "description": "Parameter table_id",
                    "type": "string"
                  },
                  "sort_columns": {
                    "description": "Parameter sort_columns",
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "ascending": {
                    "description": "Parameter ascending",
                    "type": "string"
                  },
                  "in_place": {
                    "description": "Parameter in_place",
                    "type": "boolean"
                  },
                  "new_table_name": {
                    "description": "Parameter new_table_name",
                    "type": "string"
                  }
                },
                "example": {
                  "table_id": "dt_abc12345"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful tool execution",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPToolResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MCPToolResponse": {
        "type": "object",
        "description": "MCP tool execution response",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Operation success status"
          },
          "message": {
            "type": "string",
            "description": "Human-readable message"
          },
          "error": {
            "type": "string",
            "description": "Error details if success=false"
          },
          "data": {
            "type": "object",
            "description": "Tool-specific response data"
          }
        }
      },
...
    }
  }
}